✅ 添付ファイルアップロード処理の単体テストを追加 - MockMultipartFileを使用して、画像ファイルのアップロードを検証 - Questionエンティティとの連携をテスト - 今後のテスト拡張の基盤として使用予定

+++ src/main/java/com/yuru/archive/answer/AnswerController.java
@@ -60,7 +60,7 @@ public class AnswerController {
 	
 	@PreAuthorize("isAuthenticated()")
 	@PostMapping("/create/{id}")
-	public String createAnswer(Model model, @PathVariable("id") Integer id, @Valid AnswerForm answerForm,
+	public String createAnswer(Model model, @PathVariable("id") Long id, @Valid AnswerForm answerForm,
 			BindingResult bindingResult, Principal principal) {
 		Question question = this.questionService.getQuestion(id);
 		SiteUser siteUser = this.userService.getUser(principal.getName());
